pyproject.toml
[project]
name = "leaflets"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "easyocr>=1.7.2",
    "jupyter>=1.1.1",
    "loguru>=0.7.2",
    "openai>=1.57.4",
    "pandas>=2.2.3",
    "pillow>=10.4.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "sqlalchemy>=2.0.36",
    "tqdm>=4.66.5",
    "transformers>=4.47.0",
    "ultralytics>=8.3.9",
]

# ruff settings (https://docs.astral.sh/ruff/)
[tool.ruff]
# Set the maximum line length to 120.
line-length = 120

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint]
ignore = ["E402", "W293", "F405"]

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"
